Rename api script
